requirements.txt
yfinance
duckdb
pyyaml
pandas
streamlit
plotly